Rename core module to dakota
