+++ csdms/dakota/tests/test_dakota.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# Tests for the csdms.dakota.core module.
+# Tests for the csdms.dakota.dakota module.
 #
 # Call with:
 #   $ nosetests -sv
@@ -12,7 +12,7 @@ import filecmp
 from subprocess import CalledProcessError
 from nose.tools import (raises, assert_is_instance, assert_true,
                         assert_is_none, assert_equal)
-from csdms.dakota.core import Dakota
+from csdms.dakota.dakota import Dakota
 from csdms.dakota.utils import is_dakota_installed
 from . import start_dir, data_dir
