+++ csdms/dakota/tests/test_plugin_hydrotrend_run.py
@@ -12,7 +12,7 @@ import shutil
 import filecmp
 import glob
 from nose.tools import with_setup, assert_true
-from csdms.dakota.core import Dakota
+from csdms.dakota.dakota import Dakota
 from csdms.dakota.plugins.hydrotrend import is_installed \
     as is_hydrotrend_installed
 from csdms.dakota.utils import is_dakota_installed
